Catalog service jdbc persistence reference.

+++ catalog-service/src/main/java/com/nabhatek/catalogservice/domain/ProductAlreadyExistsException.java
@@ -1,7 +1,7 @@
 package com.nabhatek.catalogservice.domain;
 
 public class ProductAlreadyExistsException extends RuntimeException {
-    public ProductAlreadyExistsException(Integer Id) {
+    public ProductAlreadyExistsException(Long Id) {
         super("Product already exists: " + Id);
     }
 }
