+++ catalog-service/src/main/java/com/nabhatek/catalogservice/domain/ProductNotFoundException.java
@@ -1,7 +1,7 @@
 package com.nabhatek.catalogservice.domain;
 
 public class ProductNotFoundException extends RuntimeException {
-    public ProductNotFoundException(Integer id) {
+    public ProductNotFoundException(Long id) {
         super("Product not found: " + id);
     }
 }
